requirements.txt
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
cached-property==1.5.1
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.3
django-crispy-forms==1.9.0
django-filter==2.2.0
google-auth==1.11.3
google-auth-oauthlib==0.4.1
googleads==23.0.1
idna==2.9
isodate==0.6.0
lxml==4.5.0
oauthlib==3.1.0
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.0
six==1.14.0
sqlparse==0.3.1
suds-jurko==0.6
urllib3==1.25.8
xmltodict==0.12.0
zeep==3.4.0
